Bump pixel dungeon to 1.9.1

+++ core/src/com/watabou/pixeldungeon/items/Heap.java
@@ -1,6 +1,6 @@
 /*
  * Pixel Dungeon
- * Copyright (C) 2012-2014  Oleg Dolya
+ * Copyright (C) 2012-2015 Oleg Dolya
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -21,16 +21,22 @@ import java.util.LinkedList;
 
 import com.badlogic.gdx.utils.reflect.ClassReflection;
 import com.watabou.noosa.audio.Sample;
+import com.watabou.noosa.tweeners.AlphaTweener;
 import com.watabou.pixeldungeon.Assets;
 import com.watabou.pixeldungeon.Badges;
 import com.watabou.pixeldungeon.Dungeon;
 import com.watabou.pixeldungeon.Statistics;
+import com.watabou.pixeldungeon.actors.buffs.Buff;
+import com.watabou.pixeldungeon.actors.buffs.Burning;
+import com.watabou.pixeldungeon.actors.buffs.Frost;
 import com.watabou.pixeldungeon.actors.hero.Hero;
+import com.watabou.pixeldungeon.actors.mobs.Mimic;
 import com.watabou.pixeldungeon.actors.mobs.Wraith;
 import com.watabou.pixeldungeon.effects.CellEmitter;
 import com.watabou.pixeldungeon.effects.Speck;
 import com.watabou.pixeldungeon.effects.Splash;
 import com.watabou.pixeldungeon.effects.particles.ElmoParticle;
+import com.watabou.pixeldungeon.effects.particles.FlameParticle;
 import com.watabou.pixeldungeon.effects.particles.ShadowParticle;
 import com.watabou.pixeldungeon.items.food.ChargrilledMeat;
 import com.watabou.pixeldungeon.items.food.FrozenCarpaccio;
@@ -39,13 +45,18 @@ import com.watabou.pixeldungeon.items.scrolls.Scroll;
 import com.watabou.pixeldungeon.plants.Plant.Seed;
 import com.watabou.pixeldungeon.sprites.ItemSprite;
 import com.watabou.pixeldungeon.sprites.ItemSpriteSheet;
+import com.watabou.pixeldungeon.utils.GLog;
 import com.watabou.utils.Bundlable;
 import com.watabou.utils.Bundle;
 import com.watabou.utils.Random;
 
 public class Heap implements Bundlable {
 
+	private static final String TXT_MIMIC = "This is a mimic!";
+	
 	private static final int SEEDS_TO_POTION = 3;
+	
+	private static final float FADE_TIME = 0.6f;
 	
 	public enum Type {
 		HEAP, 
@@ -54,7 +65,9 @@ public class Heap implements Bundlable {
 		LOCKED_CHEST, 
 		CRYSTAL_CHEST,
 		TOMB, 
-		SKELETON
+		SKELETON,
+		MIMIC,
+		HIDDEN
 	}
 	public Type type = Type.HEAP;
 	
@@ -62,7 +75,7 @@ public class Heap implements Bundlable {
 	
 	public ItemSprite sprite;
 	
-	protected LinkedList<Item> items = new LinkedList<Item>();
+	public LinkedList<Item> items = new LinkedList<Item>();
 	
 	public int image() {
 		switch (type) {
@@ -70,6 +83,7 @@ public class Heap implements Bundlable {
 		case FOR_SALE:
 			return size() > 0 ? items.peek().image() : 0;
 		case CHEST:
+		case MIMIC:
 			return ItemSpriteSheet.CHEST;
 		case LOCKED_CHEST:
 			return ItemSpriteSheet.LOCKED_CHEST;
@@ -79,6 +93,8 @@ public class Heap implements Bundlable {
 			return ItemSpriteSheet.TOMB;
 		case SKELETON:
 			return ItemSpriteSheet.BONES;
+		case HIDDEN:
+			return ItemSpriteSheet.HIDDEN;
 		default:
 			return 0;
 		}
@@ -90,6 +106,14 @@ public class Heap implements Bundlable {
 	
 	public void open( Hero hero ) {
 		switch (type) {
+		case MIMIC:
+			if (Mimic.spawnAt( pos, items ) != null) {
+				GLog.n( TXT_MIMIC );
+				destroy();
+			} else {
+				type = Type.CHEST;
+			}
+			break;
 		case TOMB:
 			Wraith.spawnAround( hero.pos );
 			break;
@@ -106,12 +130,18 @@ public class Heap implements Bundlable {
 				}
 			}
 			break;
+		case HIDDEN:
+			sprite.alpha( 0 );
+			sprite.parent.add( new AlphaTweener( sprite, 1, FADE_TIME ) );
+			break;
 		default:
 		}
 		
-		type = Type.HEAP;
-		sprite.link();
-		sprite.drop();
+		if (type != Type.MIMIC) {
+			type = Type.HEAP;
+			sprite.link();
+			sprite.drop();
+		}
 	}
 	
 	public int size() {
@@ -171,6 +201,14 @@ public class Heap implements Bundlable {
 	
 	public void burn() {
 		
+		if (type == Type.MIMIC) {
+			Mimic m = Mimic.spawnAt( pos, items );
+			if (m != null) {
+				Buff.affect( m, Burning.class ).reignite( m );
+				m.sprite.emitter().burst( FlameParticle.FACTORY, 5 );
+				destroy();
+			}
+		}
 		if (type != Type.HEAP) {
 			return;
 		}
@@ -212,6 +250,13 @@ public class Heap implements Bundlable {
 	
 	public void freeze() {
 		
+		if (type == Type.MIMIC) {
+			Mimic m = Mimic.spawnAt( pos, items );
+			if (m != null) {
+				Buff.prolong( m, Frost.class, Frost.duration( m ) * Random.Float( 1.0f, 1.5f ) );
+				destroy();
+			}
+		}
 		if (type != Type.HEAP) {
 			return;
 		}
